requirements.txt
streamlit==1.31.0
openai==1.12.0
azure-ai-formrecognizer==3.3.0
python-docx==1.0.1
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.0
pandas==2.1.1